Move the `file-store-ot` module under the `@bayou` namespace.
