+++ local-modules/@bayou/file-store-ot/Errors.js
@@ -2,7 +2,7 @@
 // Licensed AS IS and WITHOUT WARRANTY under the Apache License,
 // Version 2.0. Details: <http://www.apache.org/licenses/LICENSE-2.0>
 
-import { StoragePath } from 'file-store-ot';
+import { StoragePath } from '@bayou/file-store-ot';
 import { RevisionNumber } from '@bayou/ot-common';
 import { TString } from '@bayou/typecheck';
 import { FrozenBuffer, InfoError, UtilityClass } from '@bayou/util-common';
