+++ local-modules/@bayou/file-store-ot/tests/test_StoragePath.js
@@ -5,7 +5,7 @@
 import { assert } from 'chai';
 import { describe, it } from 'mocha';
 
-import { StoragePath } from 'file-store-ot';
+import { StoragePath } from '@bayou/file-store-ot';
 
 /** {array<string>} List of valid paths. */
 const VALID_PATHS = [
@@ -62,7 +62,7 @@ const NON_STRINGS = [
   { '/x': '/y' }
 ];
 
-describe('file-store-ot/StoragePath', () => {
+describe('@bayou/file-store-ot/StoragePath', () => {
   describe('allPrefixes()', () => {
     it('should work as expected', () => {
       function test(value, expected) {
