+++ local-modules/doc-server/BaseDataManager.js
@@ -2,7 +2,7 @@
 // Licensed AS IS and WITHOUT WARRANTY under the Apache License,
 // Version 2.0. Details: <http://www.apache.org/licenses/LICENSE-2.0>
 
-import { TransactionSpec } from 'file-store-ot';
+import { TransactionSpec } from '@bayou/file-store-ot';
 
 import BaseComplexMember from './BaseComplexMember';
 import ValidationStatus from './ValidationStatus';
