+++ local-modules/doc-server/tests/test_BaseControl.js
@@ -11,7 +11,7 @@ import { MockChange, MockDelta, MockOp, MockSnapshot } from '@bayou/ot-common/mo
 import { DurableControl, FileAccess } from 'doc-server';
 import { MockControl } from 'doc-server/mocks';
 import { MockFile } from 'file-store/mocks';
-import { Errors as fileStoreOt_Errors, TransactionSpec } from 'file-store-ot';
+import { Errors as fileStoreOt_Errors, TransactionSpec } from '@bayou/file-store-ot';
 import { Timestamp } from '@bayou/ot-common';
 import { TheModule as mocks_TheModule } from '@bayou/ot-common/mocks';
 import { Errors, FrozenBuffer } from '@bayou/util-common';
