+++ local-modules/file-store-local/LocalFile.js
@@ -7,7 +7,7 @@ import path from 'path';
 
 import { Codec } from '@bayou/codec';
 import { BaseFile } from 'file-store';
-import { FileChange, FileSnapshot } from 'file-store-ot';
+import { FileChange, FileSnapshot } from '@bayou/file-store-ot';
 import { RevisionNumber } from '@bayou/ot-common';
 import { Condition, Delay, Mutex } from '@bayou/promise-util';
 import { TString } from '@bayou/typecheck';
