+++ local-modules/file-store-local/LocalFileStore.js
@@ -8,7 +8,7 @@ import path from 'path';
 import { Codec } from '@bayou/codec';
 import { Dirs } from 'env-server';
 import { BaseFileStore, FileCache } from 'file-store';
-import { TheModule as fileStoreOt_TheModule } from 'file-store-ot';
+import { TheModule as fileStoreOt_TheModule } from '@bayou/file-store-ot';
 import { Logger } from '@bayou/see-all';
 
 import LocalFile from './LocalFile';
