+++ local-modules/file-store-local/tests/TempFiles.js
@@ -9,7 +9,7 @@ import path from 'path';
 
 import { Codec } from '@bayou/codec';
 import { LocalFile } from 'file-store-local';
-import { TheModule as fileStoreOt_TheModule } from 'file-store-ot';
+import { TheModule as fileStoreOt_TheModule } from '@bayou/file-store-ot';
 import { UtilityClass } from '@bayou/util-common';
 
 /** {Codec} Codec instance to use for `LocalFile` instances. */
