+++ local-modules/file-store-local/tests/test_LocalFile.js
@@ -5,7 +5,7 @@
 import { assert } from 'chai';
 import { describe, it } from 'mocha';
 
-import { TransactionOp, TransactionSpec } from 'file-store-ot';
+import { TransactionOp, TransactionSpec } from '@bayou/file-store-ot';
 import { FrozenBuffer } from '@bayou/util-common';
 
 import TempFiles from './TempFiles';
