+++ local-modules/file-store/BaseFile.js
@@ -2,7 +2,7 @@
 // Licensed AS IS and WITHOUT WARRANTY under the Apache License,
 // Version 2.0. Details: <http://www.apache.org/licenses/LICENSE-2.0>
 
-import { StorageId, StoragePath, TransactionSpec } from 'file-store-ot';
+import { StorageId, StoragePath, TransactionSpec } from '@bayou/file-store-ot';
 import { TBoolean, TInt, TMap, TObject, TSet } from '@bayou/typecheck';
 import { CommonBase, Errors, FrozenBuffer } from '@bayou/util-common';
