+++ local-modules/file-store/package.json
@@ -5,7 +5,7 @@
   "dependencies": {
     "@bayou/codec": "local",
     "@bayou/deps-util-server": "local",
-    "file-store-ot": "local",
+    "@bayou/file-store-ot": "local",
     "hooks-server": "local",
     "@bayou/promise-util": "local",
     "@bayou/typecheck": "local",
